+++ content-food-menu.php
@@ -18,8 +18,20 @@
  *
  * @package    Auberge
  * @copyright  2014 WebMan - Oliver Juhas
- * @version    1.0
+ *
+ * @since    1.0
+ * @version  1.1
  */
+
+
+
+/**
+ * Requirements check
+ */
+
+	if ( ! current_theme_supports( 'nova_menu_item' ) ) {
+		return;
+	}
 
 
 
@@ -40,12 +52,12 @@ $pagination_suffix = wm_paginated_suffix( 'small', 'nova_menu_item' );
 	if (
 			has_post_thumbnail()
 			&& ! $pagination_suffix
-			&& apply_filters( 'wmhook-entry-featured-image-display', true )
+			&& apply_filters( 'wmhook_entry_featured_image_display', true )
 		) :
 
 		$image_size = ( is_single() ) ? ( WM_IMAGE_SIZE_SINGULAR ) : ( WM_IMAGE_SIZE_ITEMS_MENU );
 		$image_link = ( ! is_single() && trim( strip_tags( get_the_content() ) ) ) ? ( array( get_permalink() ) ) : ( wp_get_attachment_image_src( get_post_thumbnail_id(), 'full' ) );
-		$image_link = array_filter( (array) apply_filters( 'wmhook-entry-image-link', $image_link ) );
+		$image_link = array_filter( (array) apply_filters( 'wmhook_entry_image_link', $image_link ) );
 
 		?>