+++ content.php
@@ -15,7 +15,9 @@
  *
  * @package    Auberge
  * @copyright  2014 WebMan - Oliver Juhas
- * @version    1.0
+ *
+ * @since    1.0
+ * @version  1.1
  */
 
 
@@ -34,12 +36,12 @@ $pagination_suffix = wm_paginated_suffix( 'small', 'post' );
 	if (
 			has_post_thumbnail()
 			&& ! $pagination_suffix
-			&& apply_filters( 'wmhook-entry-featured-image-display', true )
+			&& apply_filters( 'wmhook_entry_featured_image_display', true )
 		) :
 
 		$image_size = ( is_single() ) ? ( WM_IMAGE_SIZE_SINGULAR ) : ( WM_IMAGE_SIZE_ITEMS );
 		$image_link = ( is_single() ) ? ( wp_get_attachment_image_src( get_post_thumbnail_id(), 'full' ) ) : ( array( get_permalink() ) );
-		$image_link = array_filter( (array) apply_filters( 'wmhook-entry-image-link', $image_link ) );
+		$image_link = array_filter( (array) apply_filters( 'wmhook_entry_image_link', $image_link ) );
 
 		?>