+++ image.php
@@ -4,7 +4,9 @@
  *
  * @package    Auberge
  * @copyright  2014 WebMan - Oliver Juhas
- * @version    1.0
+ *
+ * @since    1.0
+ * @version  1.1
  */
 
 
@@ -28,7 +30,7 @@ get_header();
 			 */
 
 				$image_size = WM_IMAGE_SIZE_SINGULAR;
-				$image_link = array_filter( (array) apply_filters( 'wmhook-entry-image-link', wp_get_attachment_image_src( get_the_ID(), 'full' ) ) );
+				$image_link = array_filter( (array) apply_filters( 'wmhook_entry_image_link', wp_get_attachment_image_src( get_the_ID(), 'full' ) ) );
 
 				?>