+++ loop-food-menu.php
@@ -4,8 +4,20 @@
  *
  * @package    Auberge
  * @copyright  2014 WebMan - Oliver Juhas
- * @version    1.0
+ *
+ * @since    1.0
+ * @version  1.1
  */
+
+
+
+/**
+ * Requirements check
+ */
+
+	if ( ! current_theme_supports( 'nova_menu_item' ) ) {
+		return;
+	}